add refreshtoken rotation, blacklist, mybatis->jpa

+++ src/main/java/com/flab/msk_shopping/auth/jwt/service/JwtService.java
@@ -1,32 +1,39 @@
 package com.flab.msk_shopping.auth.jwt.service;
 
 import com.flab.msk_shopping.auth.jwt.JwtProvider;
-import com.flab.msk_shopping.auth.jwt.controller.dto.RefreshTokenDto;
 import com.flab.msk_shopping.auth.jwt.controller.dto.SignUpRequestDto;
 import com.flab.msk_shopping.auth.jwt.domain.JwtToken;
+import com.flab.msk_shopping.auth.jwt.domain.RefreshToken;
 import com.flab.msk_shopping.auth.jwt.domain.User;
-import com.flab.msk_shopping.auth.jwt.repository.AccessTokenRepository;
+import com.flab.msk_shopping.auth.jwt.repository.RedisRepository;
 import com.flab.msk_shopping.auth.jwt.repository.RefreshTokenRepository;
 import com.flab.msk_shopping.auth.jwt.repository.UserRepository;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
+import io.jsonwebtoken.JwtException;
+import io.jsonwebtoken.Jwts;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.net.URI;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.Base64;
 import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
 
 @Service
 public class JwtService {
 
-    @Autowired
-    AccessTokenRepository accessTokenRepository;
+    //질문 : 생성자 패턴으로 어떻게 바꾸지?
 
     @Autowired
     RefreshTokenRepository refreshTokenRepository;
@@ -37,16 +44,41 @@ public class JwtService {
     @Autowired
     JwtProvider jwtProvider;
 
+    @Autowired
+    RedisRepository redisRepository;
+
     @Value("${jwt.sign_in.password.salt}")
     private String passwordSalt;
 
 
     @Transactional
-    public void logout(Long userId) {
+    public void logout(String accessToken) {
+        if (accessToken == null) {
+            throw new IllegalStateException();
+        }
+
+        String actualToken = accessToken.replaceFirst("Bearer ", ""); // Bearer 제거
+        Claims claimsJws;
         try {
-            // 멤버의 access, refresh 토큰 삭제
-            accessTokenRepository.deleteAccessTokenByUserId(userId);
-            refreshTokenRepository.deleteRefreshTokenByUserId(userId);
+            // 토큰 읽어서 userId 알아내기
+            claimsJws = Jwts.parser()
+                    .verifyWith(JwtProvider.getSigningKey())
+                    .build()
+                    .parseSignedClaims(actualToken)
+                    .getPayload();
+        } catch (JwtException e) {
+            throw new IllegalStateException();
+        }
+        Long userId = Long.parseLong(claimsJws.getSubject());
+        try {
+            // 멤버의 refresh 토큰 삭제
+            RefreshToken refreshToken = refreshTokenRepository.findRefreshTokenByUserId(userId);
+            refreshTokenRepository.deleteByUserId(userId);
+
+            // save RT at black list
+            Long tillExpiration = refreshToken.getExpiration().getTime() - (new Date(System.currentTimeMillis())).getTime();
+            redisRepository.setData(refreshToken.getTokenValue(), String.valueOf(userId), tillExpiration);
+
         } catch(Exception e) {
             throw new IllegalArgumentException(e.getMessage());
         }
@@ -56,29 +88,29 @@ public class JwtService {
     public JwtToken login(String email, String password) {
         User user = userRepository.findUserByEmail(email);
         if(user == null) {
-//            throw new NonExistentMemberException(userId, "회원이 존재하지 않습니다.");
             throw new RuntimeException("회원이 존재하지 않습니다.");
-
         }
 
         // 비밀번호를 해싱
         String encryptedPassword = getEncryptedPassword(password);
         if(!encryptedPassword.equals(user.getPassword())) { // 같지 않음
-//            throw new IncorrectPasswordException(userId, "비밀번호가 틀립니다.");
             throw new RuntimeException("비밀번호가 틀립니다.");
-
         }
 
         // 토큰 발행
-        return jwtProvider.createJwtToken(user.getUserId(), user.getNickName(), user.getIsAdmin());
+        String accessToken = jwtProvider.createAccessToken(user.getUserId(), user.getEmail(), user.getIsAdmin());
+        String refreshToken = jwtProvider.createRefreshToken(user.getUserId());
+
+       return new JwtToken(accessToken, refreshToken);
     }
 
     // refreshToken을 읽고 accessToken을 재발급합니다.
     @Transactional
     public JwtToken reissueAccessToken(String refreshToken) {
-        Jws<Claims> claims = jwtProvider.parseClaims(refreshToken);
+        if (refreshToken == null) return null;
 
-        // 리턴 값이 null이라면 refresh token이 만료
+        Jws<Claims> claims = jwtProvider.parseClaims(refreshToken);
+        // claims가 null이라면 refresh token이 만료
         if(claims == null) {
             // refresh Token을 DB에서 삭제
             refreshTokenRepository.deleteRefreshTokenByTokenValue(refreshToken);
@@ -86,34 +118,42 @@ public class JwtService {
         }
 
         Long userId = Long.parseLong(claims.getBody().getSubject());
-        String nickname = claims.getPayload().get("nickname", String.class);
-        boolean isAdmin = Boolean.parseBoolean(claims.getPayload().get("isAdmin", String.class));
+
+        // 블랙리스트에 올라있는 토큰인지 확인
+        if(redisRepository.getData(refreshToken)!=null){
+            // todo : custom exception
+            throw new RuntimeException("블랙리스트 리프레시 토큰입니다.");
+        }
 
         // refreshToken과 DB에 저장된 refreshToken의 값이 일치하는지 확인
-        RefreshTokenDto findRefreshToken = refreshTokenRepository.findRefreshTokenByUserId(
-                userId);
+        RefreshToken findRefreshToken = refreshTokenRepository.findRefreshTokenByUserId(userId);
         if(!findRefreshToken.getTokenValue().equals(refreshToken)) {
             throw new RuntimeException("잘못된 리프레시 토큰입니다.");
         }
 
-        // 재발급
-        String accessToken = jwtProvider.createAccessToken(userId, nickname, isAdmin);
+        // RT rotation
+        refreshTokenRepository.deleteRefreshTokenByTokenValue(refreshToken);
+        User currentUser = userRepository.findUserByUserId(userId);
 
-        return new JwtToken(accessToken, refreshToken);
+        // 재발급 (refreshToken rotation)
+        String accessTokenValue = jwtProvider.createAccessToken(userId, currentUser.getEmail(), currentUser.getIsAdmin());
+        String refreshTokenValue = jwtProvider.createRefreshToken(userId);
+
+        return new JwtToken(accessTokenValue, refreshTokenValue);
     }
 
     public void signUp(SignUpRequestDto signUpRequestDto) {
         String encryptedPassword = getEncryptedPassword(signUpRequestDto.getPassword());
 
-        User user = new User(encryptedPassword
-                ,signUpRequestDto.getEmail()
-                ,signUpRequestDto.getNickname()
-                ,signUpRequestDto.getIsAdmin()
+        User user = new User(
+                signUpRequestDto.getEmail(),
+                encryptedPassword,
+                signUpRequestDto.getIsAdmin()
                 );
 
         // DB에 회원 정보 저장
         try {
-            userRepository.addUser(user);
+            userRepository.save(user);
         } catch (DataIntegrityViolationException e) {
             throw new RuntimeException("아이디 또는 이메일 또는 연락처가 기존 회원과 중복됩니다.");
         } catch (Exception e) {
